Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -5,8 +5,8 @@ char* welcome() {
 }
 
 int main() {
-    long a;
-    scanf("%ld",&a);
+    int a;
+    scanf("%d",&a);
     (a>>32);
     printf("%d",a);
 }
